# BACK-END/API/src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access Link API",
    "description": "API for Access",
    "version": "1.0.0"
  },
  "host": [
    "api.access-link.tech"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer &lt;token&gt;**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/presence/presenceDeMesCollaborateurs/{page}/{itemsParPage}": {
      "get": {
        "tags": [
          "Présence"
        ],
        "description": "Lister les présences des collaborateurs",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Numéro de la page"
          },
          {
            "name": "itemsParPage",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Nombre d'éléments par page"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filtering",
            "in": "query",
            "description": "Filtres appliqués",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des présences.",
            "schema": {
              "$ref": "#/definitions/Presence"
            }
          },
          "401": {
            "description": "Accès non autorisé."
          }
        }
      }
    },
    "/presence/presenceDeMesCollaborateurs/export/{page}/{itemsParPage}": {
      "get": {
        "tags": [
          "Présence"
        ],
        "description": "Exporter les présences des collaborateurs",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Numéro de la page"
          },
          {
            "name": "itemsParPage",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Nombre d'éléments par page"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filtering",
            "in": "query",
            "description": "Filtres appliqués",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier Excel exporté.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "file"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Accès non autorisé."
          }
        }
      }
    },
    "/presence/mesPresences": {
      "get": {
        "tags": [
          "Présence"
        ],
        "description": "Obtenir les présences d\\'un collaborateur",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des présences du collaborateur.",
            "schema": {
              "$ref": "#/definitions/Presence"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/presence/ajouterPresenceCollab": {
      "post": {
        "tags": [
          "Présence"
        ],
        "description": "Ajouter une présence pour un collaborateur",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Informations de la présence",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idCollab": {
                  "type": "string",
                  "example": "integer"
                },
                "datePres": {
                  "type": "string",
                  "example": "string"
                },
                "hdeb": {
                  "type": "string",
                  "example": "string"
                },
                "hfin": {
                  "type": "string",
                  "example": "string"
                },
                "desc": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Présence ajoutée.",
            "schema": {
              "$ref": "#/definitions/Presence"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/presence/modifierPresenceCollab/{idPres}": {
      "put": {
        "tags": [
          "Présence"
        ],
        "description": "Modifier une présence pour un collaborateur",
        "parameters": [
          {
            "name": "idPres",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID de la présence"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Informations de la présence",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idCollab": {
                  "type": "string",
                  "example": "integer"
                },
                "datePres": {
                  "type": "string",
                  "example": "string"
                },
                "hdeb": {
                  "type": "string",
                  "example": "string"
                },
                "hfin": {
                  "type": "string",
                  "example": "string"
                },
                "desc": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Présence modifiée.",
            "schema": {
              "$ref": "#/definitions/Presence"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Absence": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "datedeb": {
          "type": "string",
          "example": "2023-07-15"
        },
        "periodeDeb": {
          "type": "number",
          "example": 1
        },
        "datefin": {
          "type": "string",
          "example": "2023-07-20"
        },
        "periodeFin": {
          "type": "number",
          "example": 2
        },
        "raison": {
          "type": "string",
          "example": "conge"
        },
        "description": {
          "type": "string",
          "example": "Congé annuel"
        },
        "accepte": {
          "type": "boolean",
          "example": true
        },
        "datereponse": {
          "type": "string",
          "example": "2023-07-16"
        },
        "dateDemande": {
          "type": "string",
          "example": "2023-07-01"
        },
        "reponseDe": {
          "$ref": "#/definitions/Collaborateur"
        },
        "modifierPar": {
          "$ref": "#/definitions/Collaborateur"
        }
      }
    },
    "Access": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "macadress": {
          "type": "string",
          "example": "00:14:22:01:23:45"
        },
        "typePoint": {
          "type": "string",
          "example": "pointeuse"
        },
        "location": {
          "type": "string",
          "example": "Bureau"
        },
        "nompoint": {
          "type": "string",
          "example": "Entrée principale"
        },
        "active": {
          "type": "boolean",
          "example": true
        },
        "collabAutorise": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        },
        "serviceAutorise": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Service"
          }
        },
        "historique": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historique"
          }
        }
      }
    },
    "Collaborateur": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "prenom": {
          "type": "string",
          "example": "John"
        },
        "nom": {
          "type": "string",
          "example": "Doe"
        },
        "mail": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "motdepasse": {
          "type": "string",
          "example": "password"
        },
        "fonction": {
          "type": "string",
          "example": "Développeur"
        },
        "grade": {
          "type": "string",
          "example": "collab"
        },
        "tokens": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Token"
          }
        },
        "absences": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Absence"
          }
        },
        "incident": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Incident"
          }
        },
        "historique": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historique"
          }
        },
        "service": {
          "$ref": "#/definitions/Service"
        },
        "horairesdefault": {
          "$ref": "#/definitions/HorairesModele"
        },
        "horaire": {
          "$ref": "#/definitions/Horaire"
        },
        "actif": {
          "type": "boolean",
          "example": true
        },
        "valide": {
          "type": "boolean",
          "example": false
        },
        "date": {
          "type": "string",
          "example": "2023-07-01"
        }
      }
    },
    "Historique": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "date": {
          "type": "string",
          "example": "2023-07-15"
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "point": {
          "$ref": "#/definitions/Access"
        },
        "typeAction": {
          "type": "string",
          "example": "Entrée"
        },
        "actionAutorise": {
          "type": "boolean",
          "example": true
        },
        "statutUtilise": {
          "type": "string",
          "example": "Identité"
        }
      }
    },
    "Horaire": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "hDebLundi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinLundi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMardi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMardi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMercredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMercredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebJeudi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinJeudi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebVendredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinVendredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebSamedi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinSamedi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebDimanche": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinDimanche": {
          "type": "string",
          "example": "17:00:00"
        }
      }
    },
    "HorairesModele": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nom": {
          "type": "string",
          "example": "Modèle standard"
        },
        "hDebLundi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinLundi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMardi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMardi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebMercredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinMercredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebJeudi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinJeudi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebVendredi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinVendredi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebSamedi": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinSamedi": {
          "type": "string",
          "example": "17:00:00"
        },
        "hDebDimanche": {
          "type": "string",
          "example": "08:00:00"
        },
        "hFinDimanche": {
          "type": "string",
          "example": "17:00:00"
        },
        "collabs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        }
      }
    },
    "Incident": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "description": {
          "type": "string",
          "example": "Problème de connexion"
        },
        "ouvert": {
          "type": "boolean",
          "example": true
        },
        "modifiePar": {
          "$ref": "#/definitions/Collaborateur"
        },
        "modifieLe": {
          "type": "string",
          "example": "2023-07-15"
        },
        "creeLe": {
          "type": "string",
          "example": "2023-07-10"
        },
        "creePar": {
          "type": "string",
          "example": "John Doe"
        },
        "dateIncident": {
          "type": "string",
          "example": "2023-07-09"
        },
        "reponse": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IncidentReponse"
          }
        }
      }
    },
    "IncidentReponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "reponse": {
          "type": "string",
          "example": "Problème résolu"
        },
        "incident": {
          "$ref": "#/definitions/Incident"
        },
        "creeLe": {
          "type": "string",
          "example": "2023-07-11"
        }
      }
    },
...
    "Presence": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "datePres": {
          "type": "string",
          "example": "2023-07-15"
        },
        "hdeb": {
          "type": "string",
          "example": "08:00:00"
        },
        "hfin": {
          "type": "string",
          "example": "17:00:00"
        },
        "modifiePar": {
          "$ref": "#/definitions/Collaborateur"
        },
        "modifieLe": {
          "type": "string",
          "example": "2023-07-16"
        },
        "creeLe": {
          "type": "string",
          "example": "2023-07-15"
        },
        "creePar": {
          "type": "string",
          "example": "Admin"
        },
        "desc": {
          "type": "string",
          "example": "Présence normale"
        }
      }
    },
    "Service": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "chefservice": {
          "$ref": "#/definitions/Collaborateur"
        },
        "nomservice": {
          "type": "string",
          "example": "Informatique"
        },
        "collabs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Collaborateur"
          }
        }
      }
    },
    "Token": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "collab": {
          "$ref": "#/definitions/Collaborateur"
        },
        "type": {
          "type": "string",
          "example": "auth"
        },
        "actif": {
          "type": "boolean",
          "example": true
        },
        "token": {
          "type": "string",
          "example": "token123"
        },
        "datecreation": {
          "type": "string",
          "example": "2023-07-15"
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}